refactor: 단일 책임 원칙에 따라 TemplateService에서 생성 로직을 TemplateFactory로 분리
- AI 응답으로부터 `Template` 엔티티를 생성하고 조립하는 역할을 전담하는 `TemplateFactory` 생성
- `TemplateService`는 이 팩토리를 사용하여 `Template` 생성을 위임하고 자신은 비즈니스 흐름 제어에만 집중하도록 리팩토링
